Added exception handling
The previous code was terminating with an error as described in [this issue[(https://github.com/AustinDeric/yak/issues/18) and [this ROS Answer question](https://answers.ros.org/question/295359/tf-has-two-or-more-unconnected-trees/) when the `volume_pose` frame was not available for finding the transform. With the proposed addition, this try-catch block showcases the required error message if `volume_pose` is not available. 
~~~
[ERROR] [1530122898.477899539, 10.817000000]: "volume_pose" passed to lookupTransform argument target_frame does not exist. 
~~~

However, it as soon as the `volume_pose` frame becomes available, it allows the node to proceed instead of terminating as per the previous case. This is similar to what has been used in the [tf tutorials](http://wiki.ros.org/tf/Tutorials/tf%20and%20Time%20%28C%2B%2B%29).