Set default weight factor to 1
Hey,

I have set the default weight factor for the position constraint for the link `sensor_constraint_frame`. This is to avoid the unnecessary warning which comes up when this isn't done. 
~~~
[ WARN] [1530883644.819044064, 925.199000000]: The weight on position constraint for link 'sensor_constraint_frame' is near zero. Setting to 1.0.
~~~

The `weight` in this case, being the weighting factor for this constraint (denotes relative importance to other constraints. Closer to zero means less important) as shown in [this code](https://github.com/kunal15595/ros/blob/master/moveit/src/moveit_msgs/msg/PositionConstraint.msg).